+++ src/collective/easyform/fields.py
@@ -6,6 +6,7 @@ from collective.easyform.interfaces import IEasyFormForm
 from collective.easyform.interfaces import IFieldExtender
 from collective.easyform.interfaces import ILabel
 from collective.easyform.interfaces import IReCaptcha
+from collective.easyform.interfaces import INorobotCaptcha
 from collective.easyform.interfaces import IRichLabel
 from collective.easyform.validators import IFieldValidator
 from plone.schemaeditor.fields import FieldFactory
@@ -141,3 +142,17 @@ ReCaptchaFactory = FieldFactory(
     ReCaptcha, _(u"label_recaptcha_field", default=u"ReCaptcha")
 )
 ReCaptchaHandler = BaseHandler(ReCaptcha)
+
+
+@implementer(INorobotCaptcha)
+class NorobotCaptcha(TextLine):
+
+    """A NorobotCaptcha field
+    """
+
+
+NorobotFactory = FieldFactory(
+    NorobotCaptcha,
+    _(u'label_norobot_field', default=u'NorobotCaptcha')
+)
+NorobotCaptchaHandler = BaseHandler(NorobotCaptcha)